index: split CompositeIndex and stats types to "composite" module
Added pub(super) where needed or makes sense.
